Coupled conductance photosynthesis model

Photosynthesis model at the leaf level using the farquhar equations coupled with a conductance model. The parameters can be defined by the function f.make param and corresponds to the parameters implemented in FATES model. Note that the conductance model is selected in the param parameter.

f.A(PFD, cs, Tleaf, Tair, RH, param = f.make.param())

Arguments

PFD

Photosynthetic light at the leaf surface in micro mol m-2 s-1.

cs

CO2 concentration at the leaf surface.

Tleaf

Leaf temperature in Kelvin.

Tair

Air temperature in Kelvin.

RH

Relative humidity at the leaf surface (between 0 and 100).

param

List of parameters given by f.make.param().

Value

List of different variables: - A: Raw assimilation of the leaf in micromol.m-2.s-1.

- Ac: Rubisco limitation assimilation of the leaf in micromol.m-2.s-1.

- Aj: Electron transport rate assimilation of the leaf in micromol.m-2.s-1.

- Ap: TPU rate of the leaf in micromol.m-2.s-1.

- Ag: Gross assimilation in micromol.m-2.s-1.

- Rd: Respiration rate in micromol.m-2.s-1.

- gs: Conductance of the leaf for water vapour in mol m-2 s-1.

- ci: Intracellular CO2 concentration in micromol.mol-1.

- ds: Leaf surface to air vapour pressure deficit in Pa.

- Transp: Water transpiration in mL m-2 s-1.

References

Farquhar, G. D., von Caemmerer, S. V., & Berry, J. A. (1980). A biochemical model of photosynthetic CO 2 assimilation in leaves of C 3 species. planta, 149, 78-90.

Fates parameters: https://github.com/NGEET/fates/blob/main/parameter_files/fates_params_default.cdl.

Examples

f.A(PFD=2000,cs=400,Tleaf=273.16+29,Tair=273.16+28,RH=70,param=f.make.param())
#> $A
#> [1] 10.63467
#> 
#> $Ac
#> [1] 10.75746
#> 
#> $Aj
#> [1] 11.95638
#> 
#> $Ap
#> [1] 26.80939
#> 
#> $Ag
#> [1] 12.20061
#> 
#> $Rd
#> [1] 1.565939
#> 
#> $gs
#> [1] 0.2121079
#> 
#> $ci
#> [1] 319.6927
#> 
#> $ds
#> [1] 1359.726
#> 
#> $Transp
#> [1] 0.05139956
#>