Canopy scale GPP calculation

Generic function to calculate the GPP within a forest (Here GPP = sum of Anet at the canopy level, so it takes into account the leaf mitochondrial respiration)

f.GPP(
  meteo_hourly,
  Vcmax_Profile,
  Jmax_Profile,
  Rd_Profile,
  Tp_Profile,
  g0_Profile,
  g1_Profile,
  gsmin,
  canopy,
  Patm = 100,
  ...
)

Arguments

meteo_hourly

See f.canopy.interception doc. In addition to the requirement for f.canopy.interception, the leaf temperature has to be informed within the column Tleaf.

Vcmax_Profile

Vector of the values of Vcmax at the reference temperature at each layer of the canopy.

Jmax_Profile

Vector of the values of Jmax at the reference temperature at each layer of the canopy.

Rd_Profile

Vector of the values of Rd at the reference temperature at each layer of the canopy.

Tp_Profile

Vector of the values of Tp at the reference temperature at each layer of the canopy.

g0_Profile

Vector of the values of g0 at the reference temperature at each layer of the canopy.

g1_Profile

Vector of the values of g1 at the reference temperature at each layer of the canopy.

gsmin

Minimum stomatal conductance for water to consider. This value will be used as the minimum conductance value to avoid 0 and negative values obtained from the coupled assimilation and conductance models.

canopy

Description of the canopy interception (see canopy_interception function).

Patm

Atmospheric pressure (used to calculate the transpiration).

...

Other parameters of the photosynthetic model, without gradients, for example curvature factor, quantum yield.. see the help of f.make.param().

Examples

#See vignettes on github