Norman 1979 Radiation interception model

Converted into a R code from the original code of Gordon Bonan: Bonan, G. (2019). Climate Change and Terrestrial Ecosystem Modeling. Cambridge: Cambridge University Press. doi:10.1017/9781107339217

f.Norman.Radiation(
  Rho = 0.1,
  Tau = 0.05,
  Rho_soil_dir = 0.1,
  Rho_soil_dif = 0.1,
  cosz,
  chil,
  clumpfac,
  dLAI,
  nlayers,
  PARdir = 0.8,
  PARdif = 0.2
)

Arguments

Rho

Leaf reflectance.

Tau

Leaf transmittance.

Rho_soil_dir

Direct beam albedo of ground (soil).

Rho_soil_dif

Diffuse albedo of ground (soil).

cosz

Cosinus of the solar zenith angle.

chil

Index of departure of the leaf angles from a spherical distribution. -0.4 < chil < 0.6.

clumpfac

Clumping factor, index of non random spatial distribution of leaves. = 1 for randomly spaced leaves, <1 for clumed leaves (Chen et al. 2012).

dLAI

LAI of each one of the n layers of vegetation in the canopy, layer 1 is the top of canopy, layer n is the bottom.

nlayers

Number of vegetation layers.

PARdir

Atmospheric direct beam solar radiation (W/m2).

PARdif

Atmospheric diffuse solar radiation (W/m2).

Value

list of output: PARsun Absorbed PFD by the sunlit leaves PARsha Absorbed PFD by the shaded leaves fracsun Proportion of sunlit leaves fracsha Proportion of shaded leaves

Examples

f.Norman.Radiation(
  Rho = 0.1, Tau = 0.05, PARdir = 1000, PARdif = 200,
  dLAI = c(rep(6 / 20, 20)), nlayers = 20, Rho_soil_dif = 0.1, Rho_soil_dir = 0.1,
  cosz = 0.88, chil = 0.1, clumpfac = 0.8
)
#> [1] "Radiation model for a total LAI of  6"
#> $PARsun
#>  [1] 656.2927 636.4681 619.5413 605.0824 592.7271 582.1661 573.1369 565.4166
#>  [9] 558.8160 553.1745 548.3558 544.2444 540.7428 537.7687 535.2531 533.1387
#> [17] 531.3784 529.9342 528.7764 527.8832
#> 
#> $PARsha
#>  [1] 137.912146 118.087586 101.160730  86.701868  74.346552  63.785540
#>  [7]  54.756331  47.036052  40.435469  34.793949  29.975232  25.863870
#> [13]  22.362258  19.388143  16.872572  14.758190  12.997878  11.553661
#> [19]  10.395886   9.502641
#> 
#> $fracsha
#>  [1] 0.2564071 0.3575700 0.4449702 0.5204798 0.5857167 0.6420784 0.6907722
#>  [8] 0.7328415 0.7691874 0.8005885 0.8277177 0.8511560 0.8714057 0.8889005
#> [15] 0.9040151 0.9170735 0.9283553 0.9381023 0.9465233 0.9537986
#> 
#> $fracsun
#>  [1] 0.74359294 0.64242998 0.55502985 0.47952017 0.41428329 0.35792164
#>  [7] 0.30922778 0.26715852 0.23081263 0.19941146 0.17228230 0.14884395
#> [13] 0.12859431 0.11109954 0.09598488 0.08292650 0.07164467 0.06189769
#> [19] 0.05347675 0.04620144
#>