Calculation of the minimal conductance given by a particular coupled conductance and photosynthesis model

The minimal conductance of a model depends on the parameters of the model (ie g0 and g1) but also on the minimum A value, which corresponds to the dark respiration. Knowing the minimal conductance is important because the conductance can become negative and lead to unrealistic values in photosynthesis models

f.gsmin(
  RdRef = 0.825,
  RdHa = 46390,
  RdHd = 150650,
  RdS = 490,
  Tleaf = 300,
  cs = 400,
  ds = 1000,
  g0 = 0.02,
  g1 = 4.1,
  power = 0.5,
  model = "USO"
)

Arguments

RdRef

Respiration value at the reference temperature in micromol.m-2.s-1.

RdHa

Energie of activation for Rd in J.mol-1.

Tleaf

Leaf temperature in Kelvin

cs

CO2 concentration at the leaf surface in ppm

ds

Leaf to air vapor pressure deficit in Pa

g0

Constant of the conductance model, representing the conductance when A is 0, in mol.m-2.s-1, usually around 0.01.

g1

Slope parameter, between 1.14 and 3.58 KPa^0.5 (Wu et al., 2019).

power

Power of VPDl in USO model. By default power=0.5 as in Medlyn article.

RH

Air relative humidity (0 to 100), only necessary for the BWB stomatal conductance model

Value

Minimum conductance in mol m-2 s-1

Examples

f.gsmin(RdRef=  0.825,RdHa=  46390,RdHd=150650,RdS=490,Tleaf=300,cs=400,ds=1000,g0=0.02,g1=4.1,power=0.5,model="USO")
#> [1] 0.009196801